Broke CommentStateUtil.java out of CommentAdapter.java
